release/3.4.0-beta.1: updated Appodeal Android SDK part;
