+++ .gitignore
@@ -49,4 +49,7 @@ local.properties
 buck-out/
 \.buckd/
 *.keystore
+
+package-lock.json
+yarn.lock
