+++ README.md
@@ -193,6 +193,7 @@ allprojects {
         ...
         jcenter()
         maven { url "https://artifactory.appodeal.com/appodeal" }
+        maven { url "https://artifactory.appodeal.com/appodeal-beta-public" }
         ...
     }
 }
