+++ android/src/main/java/com/appodeal/rnappodeal/RNAppodealPackage.java
@@ -2,37 +2,32 @@ package com.appodeal.rnappodeal;
 
 import androidx.annotation.NonNull;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
 import com.facebook.react.ReactPackage;
+import com.facebook.react.bridge.JavaScriptModule;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
-import com.facebook.react.bridge.JavaScriptModule;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 
 
 public class RNAppodealPackage implements ReactPackage {
     @NonNull
     @Override
-    public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-      return Arrays.<NativeModule>asList(
-              new RNAppodealModule(reactContext)
-      );
+    public List<NativeModule> createNativeModules(@NonNull ReactApplicationContext reactContext) {
+        return Collections.singletonList(new RNAppodealModule(reactContext));
     }
 
     // Deprecated from RN 0.47
     public List<Class<? extends JavaScriptModule>> createJSModules() {
-      return Collections.emptyList();
+        return Collections.emptyList();
     }
 
     @NonNull
     @Override
-    public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-      return Arrays.<ViewManager>asList(
-              new RNAppodealBannerManager(),
-              new RNAppodealMrecManager()
-      );
+    public List<ViewManager> createViewManagers(@NonNull ReactApplicationContext reactContext) {
+        return Arrays.asList(new RNAppodealBannerManager(), new RNAppodealMrecManager());
     }
 }
