+++ example/android/build.gradle
@@ -1,12 +1,12 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-def REACT_NATIVE_VERSION = new File(['node', '--print',"JSON.parse(require('fs').readFileSync(require.resolve('react-native/package.json'), 'utf-8')).version"].execute(null, rootDir).text.trim())
+def REACT_NATIVE_VERSION = new File(['node', '--print', "JSON.parse(require('fs').readFileSync(require.resolve('react-native/package.json'), 'utf-8')).version"].execute(null, rootDir).text.trim())
 
 buildscript {
     ext {
         minSdkVersion = 21
-        compileSdkVersion = 33
-        targetSdkVersion = 33
+        compileSdkVersion = 34
+        targetSdkVersion = 34
         ndkVersion = "21.4.7075529"
     }
     repositories {
@@ -14,25 +14,26 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath("com.android.tools.build:gradle:4.2.2")
+        classpath('com.google.gms:google-services:4.4.2')
+        classpath('com.android.tools.build:gradle:7.4.2')
     }
 }
 
 allprojects {
     repositories {
         google()
-        jcenter()
-        mavenLocal()
         mavenCentral()
         // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
         maven { url("$rootDir/../node_modules/react-native/android") }
         // Android JSC is installed from npm
         maven { url("$rootDir/../node_modules/jsc-android/dist") }
         maven { url 'https://www.jitpack.io' }
+        // Appodeal SDK
         maven { url "https://artifactory.appodeal.com/appodeal" }
+        maven { url "https://artifactory.appodeal.com/appodeal-beta-public" }
     }
 
-    configurations.all {
+    configurations.configureEach {
         resolutionStrategy {
             // Remove this override in 0.65+, as a proper fix is included in react-native itself.
             force "com.facebook.react:react-native:" + REACT_NATIVE_VERSION
