+++ example/package.json
@@ -7,40 +7,46 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint ."
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@react-native-community/segmented-control": "^2.2.1",
-    "@react-navigation/native": "^6.0.10",
-    "@react-navigation/native-stack": "^6.6.2",
-    "@types/react-native-vector-icons": "^6.4.10",
+    "@react-native-segmented-control/segmented-control": "^2.5.4",
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "react": "18.1.0",
-    "react-native": "0.70.3",
+    "react-native": "0.70.10",
     "react-native-appodeal": "../",
-    "react-native-safe-area-context": "^4.2.5",
-    "react-native-screens": "3.18.2",
-    "react-native-vector-icons": "^9.1.0"
+    "react-native-safe-area-context": "^4.11.0",
+    "react-native-screens": "^3.34.0",
+    "react-native-vector-icons": "^10.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
-    "@types/jest": "^29.1.0",
+    "@tsconfig/react-native": "^2.0.2",
+    "@types/jest": "^26.0.23",
     "@types/react": "^18.0.21",
-    "@types/react-native": "^0.70.0",
-    "@types/react-test-renderer": "^17.0.1",
+    "@types/react-native": "^0.70.6",
+    "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.37.0",
+    "@typescript-eslint/parser": "^5.37.0",
     "babel-jest": "^26.6.3",
-    "eslint": "^7.14.0",
+    "eslint": "^7.32.0",
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "0.72.3",
     "react-test-renderer": "18.1.0",
-    "typescript": "^3.8.3"
-  },
-  "resolutions": {
-    "@types/react": "18.0.21",
-    "@types/jest": "29.1.2"
+    "typescript": "^5.5.4"
   },
   "jest": {
-    "preset": "react-native"
+    "preset": "react-native",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node"
+    ]
   }
 }
