+++ example/src/components/controls/BannerSegmentedControl.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {BannerShowStyle} from '../../advertising';
 import {View, Text} from 'react-native';
 import {styles} from '../../styles';
-import SegmentedControl from '@react-native-community/segmented-control';
+import SegmentedControl from '@react-native-segmented-control/segmented-control';
 
 interface BannerSegmentedControlProps {
   visible: boolean;
